pyproject.toml
[tool.poetry]
name = "web note"
version = "1.0.0"
description = ""
authors = ["BRANYA <tbd2504@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "4.2.11"
django-ckeditor = "^6.7.1"
django-split-settings = "^1.3.1"
gunicorn = "^22.0.0"
python-dotenv = "^1.0.1"
django-bootstrap5 = "^24.2"
django-filter = "^24.2"


[tool.poetry.group.dev.dependencies]
django-debug-toolbar = "^4.3.0"
faker = "^24.14.0"
mypy = "^1.10.0"
ruff = "^0.4.2"
selenium = "^4.20.0"
webdriver-manager = "^4.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 120

[tool.ruff.format]
quote-style = "single"

[tool.mypy]
ignore_missing_imports = true
no_implicit_optional = false